requirements.txt
gradio>=4.0.0
distilabel>=1.0.0
anthropic>=0.34.0
claude-agent-sdk>=0.1.0
python-dotenv>=1.0.0
